# Conflicts:
#	.idea/runConfigurations/Tomcat_9_0_105.xml

+++ src/main/java/org/spring/example/mapper/BoardMapper.java
@@ -9,7 +9,7 @@ import java.util.List;
 public interface BoardMapper {
     List<BoardDto> selectAllBoards();
 
-    BoardDto selectById(int boardId);
+    BoardDto selectById(long boardId);
 
     int insert(BoardDto boardDto);
